# Conflicts:
#	codeblocks/config.h
#	codeblocks/doom.cbp
#	codeblocks/game-res.rc
#	codeblocks/heretic.cbp
#	codeblocks/libopl.cbp
#	codeblocks/libpcsound.cbp
#	codeblocks/libtextscreen.cbp
#	codeblocks/server.cbp
#	codeblocks/setup-res.rc
#	codeblocks/setup.cbp
#	data/doom.png
#	msvc/chocolate.sln
#	msvc/config.h
#	msvc/doom.vcproj
#	msvc/win32.rc
#	src/Makefile.am
#	src/d_iwad.h
#	src/d_mode.c
#	src/doom/d_main.c
#	src/doom/d_net.c
#	src/doom/doomdata.h
#	src/doom/g_game.c
#	src/doom/hu_stuff.c
#	src/doom/p_enemy.c
#	src/doom/p_inter.c
#	src/doom/p_map.c
#	src/doom/p_setup.c
#	src/doom/p_sight.c
#	src/doom/p_spec.c
#	src/doom/r_plane.c
#	src/doom/r_segs.c
#	src/doomtype.h
#	src/heretic/r_plane.c
#	src/hexen/h2_main.c
#	src/i_main.c
#	src/i_sdlmusic.c
#	src/i_system.c
#	src/i_video.c
#	src/i_video.h
#	src/icon.c
#	src/m_argv.c
#	src/m_config.c
#	src/m_misc.c
#	src/m_misc.h
#	src/net_client.c
#	src/net_dedicated.c
#	src/setup/display.c
#	src/setup/multiplayer.c
#	src/setup/sound.c
#	src/strife/m_saves.c
#	src/w_main.c

+++ src/heretic/p_local.h
@@ -241,8 +241,8 @@ void P_RadiusAttack(mobj_t * spot, mobj_t * source, int damage);
 // ***** P_SETUP *****
 
 extern byte *rejectmatrix;      // for fast sight rejection
-extern short *blockmaplump;     // offsets in blockmap are from here
-extern short *blockmap;
+extern int32_t *blockmaplump;   // offsets in blockmap are from here // [crispy] BLOCKMAP limit
+extern int32_t *blockmap;       // [crispy] BLOCKMAP limit
 extern int bmapwidth, bmapheight;       // in mapblocks
 extern fixed_t bmaporgx, bmaporgy;      // origin of block map
 extern mobj_t **blocklinks;     // for thing chains
